Allow not to create a database.

+++ nohttp/src/main/java/com/yanzhenjie/nohttp/cookie/DBCookieStore.java
@@ -38,43 +38,42 @@ import java.util.concurrent.locks.ReentrantLock;
  *
  * @author Yan Zhenjie.
  */
-public class DBCookieStore implements CookieStore {
+public class DBCookieStore
+  extends BasicStore<DBCookieStore> {
 
-    /**
-     * Cookie max count in disk.
-     */
     private final static int MAX_COOKIE_SIZE = 8888;
-    /**
-     * Database sync lock.
-     */
-    private Lock mLock;
-    /**
-     * Cookie entity dao.
-     */
-    private BaseDao<CookieEntity> mCookieEntityDao;
-    /**
-     * When Add and remove cookie notify.
-     */
-    private CookieStoreListener mCookieStoreListener;
 
-    private boolean mEnable = true;
+    private Context mContext;
+    private Lock mLock;
+    private BaseDao<CookieEntity> mCookieEntityDao;
+    private CookieStoreListener mCookieStoreListener;
 
     /**
      * @param context {@link Context}.
      */
     public DBCookieStore(Context context) {
-        mLock = new ReentrantLock();
-        mCookieEntityDao = new CookieEntityDao(context);
+        this.mContext = context;
+        this.mLock = new ReentrantLock();
+    }
 
-        // Delete temp cookie.
-        Where where = new Where(CookieSQLHelper.EXPIRY, Options.EQUAL, -1L);
-        mCookieEntityDao.delete(where.get());
+    private void checkInitialization() {
+        mLock.lock();
+
+        if (isEnable() && mCookieEntityDao == null) {
+            mCookieEntityDao = new CookieEntityDao(mContext);
+            // Delete temp cookie.
+            Where where = new Where(CookieSQLHelper.EXPIRY, Options.EQUAL, -1L);
+            mCookieEntityDao.delete(where.get());
+        }
+
+        mLock.unlock();
     }
 
     /**
      * The callback when adding and deleting cookies.
      *
      * @param cookieStoreListener {@link CookieStoreListener}.
+     *
      * @return {@link CookieStore}.
      */
     public CookieStore setCookieStoreListener(CookieStoreListener cookieStoreListener) {
@@ -82,19 +81,15 @@ public class DBCookieStore implements CookieStore {
         return this;
     }
 
-    public CookieStore setEnable(boolean enable) {
-        this.mEnable = enable;
-        return this;
-    }
-
     @Override
     public void add(URI uri, HttpCookie cookie) {
+        checkInitialization();
+
         mLock.lock();
         try {
-            if (mEnable && uri != null && cookie != null) {
+            if (isEnable() && uri != null && cookie != null) {
                 uri = getEffectiveURI(uri);
-                if (mCookieStoreListener != null)
-                    mCookieStoreListener.onSaveCookie(uri, cookie);
+                if (mCookieStoreListener != null) mCookieStoreListener.onSaveCookie(uri, cookie);
                 mCookieEntityDao.replace(new CookieEntity(uri, cookie));
                 trimSize();
             }
@@ -105,17 +100,20 @@ public class DBCookieStore implements CookieStore {
 
     @Override
     public List<HttpCookie> get(URI uri) {
+        checkInitialization();
+
         mLock.lock();
         try {
-            if (uri == null || !mEnable) return Collections.emptyList();
+            if (uri == null || !isEnable()) return Collections.emptyList();
 
             uri = getEffectiveURI(uri);
             Where where = new Where();
 
             String host = uri.getHost();
             if (!TextUtils.isEmpty(host)) {
-                Where subWhere = new Where(CookieSQLHelper.DOMAIN, Options.EQUAL, host).or(CookieSQLHelper.DOMAIN,
-                        Options.EQUAL, "." + host);
+                Where subWhere =
+                  new Where(CookieSQLHelper.DOMAIN, Options.EQUAL, host).or(CookieSQLHelper.DOMAIN,
+                                                                            Options.EQUAL, "." + host);
 
                 int firstDot = host.indexOf(".");
                 int lastDot = host.lastIndexOf(".");
@@ -130,8 +128,9 @@ public class DBCookieStore implements CookieStore {
 
             String path = uri.getPath();
             if (!TextUtils.isEmpty(path)) {
-                Where subWhere = new Where(CookieSQLHelper.PATH, Options.EQUAL, path).or(CookieSQLHelper.PATH,
-                        Options.EQUAL, "/").orNull(CookieSQLHelper.PATH);
+                Where subWhere =
+                  new Where(CookieSQLHelper.PATH, Options.EQUAL, path).or(CookieSQLHelper.PATH, Options.EQUAL,
+                                                                          "/").orNull(CookieSQLHelper.PATH);
                 int lastSplit = path.lastIndexOf("/");
                 while (lastSplit > 0) {
                     path = path.substring(0, lastSplit);
@@ -146,9 +145,9 @@ public class DBCookieStore implements CookieStore {
 
             List<CookieEntity> cookieList = mCookieEntityDao.getList(where.get(), null, null, null);
             List<HttpCookie> returnedCookies = new ArrayList<>();
-            for (CookieEntity cookieEntity : cookieList)
-                if (!cookieEntity.isExpired())
-                    returnedCookies.add(cookieEntity.toHttpCookie());
+            for (CookieEntity cookieEntity : cookieList) {
+                if (!cookieEntity.isExpired()) returnedCookies.add(cookieEntity.toHttpCookie());
+            }
             return returnedCookies;
         } finally {
             mLock.unlock();
@@ -157,14 +156,16 @@ public class DBCookieStore implements CookieStore {
 
     @Override
     public List<HttpCookie> getCookies() {
+        checkInitialization();
+
         mLock.lock();
         try {
-            if (!mEnable) return Collections.emptyList();
+            if (!isEnable()) return Collections.emptyList();
             List<HttpCookie> rt = new ArrayList<>();
             List<CookieEntity> cookieEntityList = mCookieEntityDao.getAll();
-            for (CookieEntity cookieEntity : cookieEntityList)
-                if (!cookieEntity.isExpired())
-                    rt.add(cookieEntity.toHttpCookie());
+            for (CookieEntity cookieEntity : cookieEntityList) {
+                if (!cookieEntity.isExpired()) rt.add(cookieEntity.toHttpCookie());
+            }
             return rt;
         } finally {
             mLock.unlock();
@@ -173,20 +174,22 @@ public class DBCookieStore implements CookieStore {
 
     @Override
     public List<URI> getURIs() {
+        checkInitialization();
+
         mLock.lock();
         try {
-            if (!mEnable) return Collections.emptyList();
+            if (!isEnable()) return Collections.emptyList();
+
             List<URI> uris = new ArrayList<>();
             List<CookieEntity> uriList = mCookieEntityDao.getAll();
             for (CookieEntity cookie : uriList) {
                 String uri = cookie.getUri();
-                if (!TextUtils.isEmpty(uri))
-                    try {
-                        uris.add(new URI(uri));
-                    } catch (Throwable e) {
-                        Logger.w(e);
-                        mCookieEntityDao.delete(CookieSQLHelper.URI + '=' + uri);
-                    }
+                if (!TextUtils.isEmpty(uri)) try {
+                    uris.add(new URI(uri));
+                } catch (Throwable e) {
+                    Logger.w(e);
+                    mCookieEntityDao.delete(CookieSQLHelper.URI + '=' + uri);
+                }
             }
             return uris;
         } finally {
@@ -196,17 +199,16 @@ public class DBCookieStore implements CookieStore {
 
     @Override
     public boolean remove(URI uri, HttpCookie httpCookie) {
+        checkInitialization();
+
         mLock.lock();
         try {
-            if (httpCookie == null || !mEnable)
-                return true;
-            if (mCookieStoreListener != null)
-                mCookieStoreListener.onRemoveCookie(uri, httpCookie);
+            if (httpCookie == null || !isEnable()) return true;
+            if (mCookieStoreListener != null) mCookieStoreListener.onRemoveCookie(uri, httpCookie);
             Where where = new Where(CookieSQLHelper.NAME, Options.EQUAL, httpCookie.getName());
 
             String domain = httpCookie.getDomain();
-            if (!TextUtils.isEmpty(domain))
-                where.and(CookieSQLHelper.DOMAIN, Options.EQUAL, domain);
+            if (!TextUtils.isEmpty(domain)) where.and(CookieSQLHelper.DOMAIN, Options.EQUAL, domain);
 
             String path = httpCookie.getPath();
             if (!TextUtils.isEmpty(path)) {
@@ -223,9 +225,11 @@ public class DBCookieStore implements CookieStore {
 
     @Override
     public boolean removeAll() {
+        checkInitialization();
+
         mLock.lock();
         try {
-            if (!mEnable) return true;
+            if (!isEnable()) return true;
             return mCookieEntityDao.deleteAll();
         } finally {
             mLock.unlock();
@@ -238,10 +242,9 @@ public class DBCookieStore implements CookieStore {
     private void trimSize() {
         int count = mCookieEntityDao.count();
         if (count > MAX_COOKIE_SIZE + 10) {
-            List<CookieEntity> rmList = mCookieEntityDao.getList(null, null, Integer.toString(count -
-                    MAX_COOKIE_SIZE), null);
-            if (rmList != null)
-                mCookieEntityDao.delete(rmList);
+            List<CookieEntity> rmList =
+              mCookieEntityDao.getList(null, null, Integer.toString(count - MAX_COOKIE_SIZE), null);
+            if (rmList != null) mCookieEntityDao.delete(rmList);
         }
     }
 
@@ -261,10 +264,11 @@ public class DBCookieStore implements CookieStore {
     }
 
     public interface CookieStoreListener {
+
         /**
          * When saving a Cookie callback.
          *
-         * @param uri    cookie corresponding uri.
+         * @param uri cookie corresponding uri.
          * @param cookie {@link HttpCookie}.
          */
         void onSaveCookie(URI uri, HttpCookie cookie);
@@ -272,7 +276,7 @@ public class DBCookieStore implements CookieStore {
         /**
          * The callback when deleting cookies.
          *
-         * @param uri    cookie corresponding uri.
+         * @param uri cookie corresponding uri.
          * @param cookie {@link HttpCookie}.
          */
         void onRemoveCookie(URI uri, HttpCookie cookie);
